remove titlebar, move Home component to containers
