--- src/renderer/components/TitleBar/TitleBar.scss
@@ -1,73 +1,0 @@
-@import "../../styles/mixins";
-@import "../../styles/variables";
-
-.title-bar {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  height: 31px;
-  background-color: darken($brand-color-wine-red, 5%);
-  visibility: visible;
-  user-select: none;
-  -webkit-app-region: drag;
-  cursor: default;
-
-  &.hidden {
-    display: none;
-  }
-
-  .title {
-    flex: 1 1 0%;
-    padding-left: 12px;
-    font-size: 12px;
-    color: #fff;
-    user-select: none;
-    cursor: default;
-  }
-
-  .window-controls {
-    display: flex;
-    height: 100%;
-    user-select: none;
-    cursor: default;
-
-    &__button {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      width: 46px;
-      height: 100%;
-      line-height: 0;
-      user-select: none;
-      -webkit-app-region: no-drag;
-      cursor: default;
-
-      &.minimize,
-      &.maximize,
-      &.restore,
-      &.toggle-menu {
-        &:hover {
-          transition: background-color 0.1s;
-          // Default Windows Minimize/Maximize button hover color.
-          background-color: rgba(255, 255, 255, .13);
-        }
-
-        &:active {
-          background-color: rgba(255, 255, 255, .23);
-        }
-      }
-
-      &.close {
-        &:hover {
-          transition: background-color 0.1s;
-          // Default Windows close button hover color.
-          background-color: #e81123;
-        }
-
-        &:active {
-          background-color: #f1707a;
-        }
-      }
-    }
-  }
-}
